+++ src/commands/utils.ts
@@ -183,7 +183,7 @@ export function addBaseOptions(command: Command): Command {
       "-w, --width <px>",
       "Chart width in pixels",
       (it: string) => parseInt(it),
-      1400,
+      1000,
     )
     .option(
       "-h, --height <px>",